Split list and multi into separate modules
